chore: move the sql pairs cleaner into sql pairs pipeline
